# use bit count intrinsics for ipv4