Most CPUs will have intrinsics for clz and similar. golang exposes these in the bits package.

I didn't do the ipv6 portion.

```
Old:
BenchmarkSummarizeRange/ipv4
BenchmarkSummarizeRange/ipv4/all
BenchmarkSummarizeRange/ipv4/all-4     	 4653746	       231 ns/op	      76 B/op
BenchmarkSummarizeRange/ipv4/32
BenchmarkSummarizeRange/ipv4/32-4      	  181884	      6143 ns/op	    2680 B/op

New
BenchmarkSummarizeRange/ipv4
BenchmarkSummarizeRange/ipv4/all
BenchmarkSummarizeRange/ipv4/all-4     	 5277212	       195 ns/op	      76 B/op
BenchmarkSummarizeRange/ipv4/32
BenchmarkSummarizeRange/ipv4/32-4      	  283497	      4066 ns/op	    2680 B/op
```